+++ R/Andromeda.R
@@ -18,9 +18,11 @@
 #' @aliases
 #' NULL Andromeda-package
 #' @importFrom methods slotNames
-#' @importFrom utils head
+#' @importFrom utils head setTxtProgressBar txtProgressBar
 #' @importFrom tidyselect all_of
 #' @import dplyr
 #' 
 "_PACKAGE"
 
+# Used to remember when a warning has already been thrown for a temp file location:
+andromedaGlobalEnv <- new.env(parent = emptyenv())