+++ NEWS.md
@@ -1,19 +1,49 @@
+Andromeda 0.3.1
+===============
+
+Bug fixes
+
+- Switching to low-level Java calls to avoid error in CRAN's Debian test environment.
+
+
+Andromeda 0.3.0
+===============
+
+Changes
+
+- Throw more informative error when user tries to use invalid Andromeda object.
+
+- Added function `getAndromedaTempDiskSpace`. Requires rJava to be installed. (Returns `NA` if not installed.)
+
+- Throw warning when disk space becomes low. Threshold defaults to 10GB, but can be altered using `options(warnDiskSpaceThreshold = <n>)`.
+
+- Allow assigning query result to an Andromeda table where the query uses that Andromeda table. (e.g. `andromeda$cars <- andromeda$cars %>% filter(speed > 10)`)
+
+- Added `progressBar` arguments to `batchApply` and `groupApply`.
+
+Bug fixes
+
+- Unzipping now also done in Andromeda temp folder (instead of system temp folder).
+
+- Lower but faster compression level now also used when saving without maintaining connection.
+
+
 Andromeda 0.2.0
 ===============
 
 Changes
 
-- Dropping nrow and ncol support, as this seems to cause instability and is not consistent with dplyr.
+- Dropping `nrow` and `ncol` support, as this seems to cause instability and is not consistent with `dplyr`.
 
-- Dropping isSorted function, as database queries are only guaranteed to stay sorted if explicitly required to (using 'arrange').
+- Dropping `isSorted` function, as database queries are only guaranteed to stay sorted if explicitly required to (using `arrange`).
 
-- Added restoreDate and restorePosixct functions.
+- Added `restoreDate` and `restorePosixct` functions.
 
 Bug fixes
 
-- Fixed typo in isAndromeda function name.
+- Fixed typo in `isAndromeda` function name.
 
-- Fixed incompatibility issue with dplyr 1.0.0 causing 'method not supported' error.
+- Fixed incompatibility issue with `dplyr` 1.0.0 causing 'method not supported' error.
 
 - Correctly handling differences in column order when appending.
 
@@ -35,7 +65,7 @@ Andromeda 0.1.2
 
 Changes
 
-- Adding fast isSorted function.
+- Adding fast `isSorted` function.
 
 - Changes in documentation for CRAN.